Zmdernar/wildcard questions (#15)
* Adds wildcard canned questions to kbs

currently wip: has a few bugs with the button interactions

* Handles question click separately

May still need to look into the networking of this implementation

* Finishes up design for wildcard canned questions v1

Pending on if we change the results page searchcontainer state,
this will be done if we are not collapsing the header.

* Minimal code refactor

Some overlapping css stuff. Oops.

* removes csstransitiongroup from searchcontainer

it wasn't used anymore so a warning was appearing. oops.

* Removed unnecessary <span>'s and updated tests

SearchContainer's tests now match the new behavior